backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-multipart
PyJWT
argon2-cffi
python-dotenv